Implements 'includeKanji' Filter option efficiently.
Adds a stage of setting up kanji to reading and reading to kanji dicts, which
makes it possible to implement the includeKanji option on the Filter function
class much more efficiently. Will also be useful for click highlighting and
filtering nodes by a level of distance away from a specified node.