Adds charge input. Fixes update/enter/select, and some other bugs.
Fixes 'ticked' function so it works properly and extracts it to outer scope from
inside update function. Refactores update function so the update/enter/merge/
exit pattern is properly implemented for nodes and links. Has simulation
restart properly in the update function by calling simulation.alpha(1)
.restart().
Updates filter function so that it 1) builds a list of the filters that have
been applied so far (so that we can implement 'unfiltering' for users - e.g.
when they filter out kun readings, and then want to put them back, we need to
know what other filters were already applied, as they may not always do apply
filters in the same order.)
Also updates filter function so that when it removes links it can deal with
both links that have been modified by a data join and those that haven't.