Changes filter mechanism.
Filters are now recorded on the 'filters' object by true/false for binary
filters, or other types, such as a string for the kanji currently filtered for.
The 'filters' object must be updated manually in the code whenever a filter is
applied to the data whether manually applied or applied by user input.