Dragging. KanjiInput box. new CharFilter function. Data preprocess. Outlines.
Adds text shadows to all text and new styling. Adds a function to preprocess
the data to remove nodes of type 'undefined', also applies this to the
'OriginalData' constant. Changes the unfilter function slightly so that it also
processes the links.
Creates new charFilter function class. This has a process method that takes
a string (usually form the filter.kanji) value which it uses to filter the
nodes.
Adds an input box to input kanji, not currently working properly because of bugs
with the unfilter/reapply filter functions I believe.
Finally adds drag functionality to all node text elements.